+++ packages/oidc-client/src/client/index.ts
@@ -1,5 +1,5 @@
 export { TailorAuthProvider } from "@client/provider";
 export { useAuth, usePlatform, useSession } from "@client/hooks";
 export type { UserInfo } from "@client/hooks";
-export type { ErrorResponse, Session } from "@lib/types";
-export { Config } from "@/lib/config";
+
+
