+++ packages/oidc-client/src/core/config.ts
