+++ packages/oidc-client/src/core/index.ts
@@ -1,0 +1,2 @@
+export type { ErrorResponse, Session } from "@core/types";
+export { Config } from "@core/config";
